pyproject.toml
[project]
name = "splitwise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.3",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
]
